Initial commit
An implementation of a Store for Tomcat session persistence. It uses
JDBC to persist to a relational database, but also compresses the
serialised objects using zstd.

This is probably a terrible idea.

The context and constraints which drove this approach were various.

We have a legacy application which ran as a single server and stored all
sessions in memory.

The sessions contained a couple of MB of data. Thus, unable to use
signed encrypted cookies to store this data.

The application was being rehoused, and moved into a cloud environment.
Continuous delivery was also being introduced. So we now have a cluster
of application instances, and we want to do zero-downtime deployment
during a normal working day.

So we needed to:

* externalise the session so that users don't lose their work and have
  a crap experience
* avoid sticky sessions so that we don't have extended deployment
  windows as sessions drain from older versions of the application
* ensure that Servlet session lifecycle events continue to happen, since
  the application relies on these to do various cleanup tasks when a
  users session expires/times out

It's the last constraint that means we need to use JDBC session
persistence. Apparently implementations which use Redis and other
options don't support that.

But we have large sessions. Reading and writing a couple of MB of data
*for every request* isn't fast.

Hence this attempt to compress the data, and make it a bit faster.